Add pending_bus_request flag in Cache
Add flag to check if there is already a request in the bus. Cache
cannot process any read/write instruction from processor if there is
already a request.
Set flag = true when enqueue
Set flag = false when cache checks that bus has completed the
transaction that was in the bus.